+++ x/feemarket/types/msg_test.go
@@ -1,0 +1,53 @@
+package types
+
+import (
+	"testing"
+
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	"github.com/stretchr/testify/suite"
+)
+
+type MsgsTestSuite struct {
+	suite.Suite
+}
+
+func TestMsgsTestSuite(t *testing.T) {
+	suite.Run(t, new(MsgsTestSuite))
+}
+
+func (suite *MsgsTestSuite) TestMsgUpdateValidateBasic() {
+	testCases := []struct {
+		name      string
+		msgUpdate *MsgUpdateParams
+		expPass   bool
+	}{
+		{
+			"fail - invalid authority address",
+			&MsgUpdateParams{
+				Authority: "invalid",
+				Params:    DefaultParams(),
+			},
+			false,
+		},
+		{
+			"pass - valid msg",
+			&MsgUpdateParams{
+				Authority: authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+				Params:    DefaultParams(),
+			},
+			true,
+		},
+	}
+
+	for _, tc := range testCases {
+		suite.Run(tc.name, func() {
+			err := tc.msgUpdate.ValidateBasic()
+			if tc.expPass {
+				suite.NoError(err)
+			} else {
+				suite.Error(err)
+			}
+		})
+	}
+}