+++ x/feemarket/types/params.go
@@ -20,34 +20,21 @@ import (
 
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/ethereum/go-ethereum/params"
 )
 
-var _ paramtypes.ParamSet = &Params{}
-
-// Parameter keys
-var (
-	ParamStoreKeyNoBaseFee                = []byte("NoBaseFee")
-	ParamStoreKeyBaseFeeChangeDenominator = []byte("BaseFeeChangeDenominator")
-	ParamStoreKeyElasticityMultiplier     = []byte("ElasticityMultiplier")
-	ParamStoreKeyBaseFee                  = []byte("BaseFee")
-	ParamStoreKeyEnableHeight             = []byte("EnableHeight")
-	ParamStoreKeyMinGasPrice              = []byte("MinGasPrice")
-	ParamStoreKeyMinGasMultiplier         = []byte("MinGasMultiplier")
-)
+var ParamsKey = []byte("Params")
 
 var (
 	// DefaultMinGasMultiplier is 0.5 or 50%
 	DefaultMinGasMultiplier = sdk.NewDecWithPrec(50, 2)
 	// DefaultMinGasPrice is 0 (i.e disabled)
 	DefaultMinGasPrice = sdk.ZeroDec()
+	// DefaultEnableHeight is 0 (i.e disabled)
+	DefaultEnableHeight = int64(0)
+	// DefaultNoBaseFee is false
+	DefaultNoBaseFee = false
 )
-
-// ParamKeyTable returns the parameter key table.
-func ParamKeyTable() paramtypes.KeyTable {
-	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
-}
 
 // NewParams creates a new Params instance
 func NewParams(
@@ -73,26 +60,13 @@ func NewParams(
 // DefaultParams returns default evm parameters
 func DefaultParams() Params {
 	return Params{
-		NoBaseFee:                false,
+		NoBaseFee:                DefaultNoBaseFee,
 		BaseFeeChangeDenominator: params.BaseFeeChangeDenominator,
 		ElasticityMultiplier:     params.ElasticityMultiplier,
 		BaseFee:                  sdkmath.NewIntFromUint64(params.InitialBaseFee),
-		EnableHeight:             0,
+		EnableHeight:             DefaultEnableHeight,
 		MinGasPrice:              DefaultMinGasPrice,
 		MinGasMultiplier:         DefaultMinGasMultiplier,
-	}
-}
-
-// ParamSetPairs returns the parameter set pairs.
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(ParamStoreKeyNoBaseFee, &p.NoBaseFee, validateBool),
-		paramtypes.NewParamSetPair(ParamStoreKeyBaseFeeChangeDenominator, &p.BaseFeeChangeDenominator, validateBaseFeeChangeDenominator),
-		paramtypes.NewParamSetPair(ParamStoreKeyElasticityMultiplier, &p.ElasticityMultiplier, validateElasticityMultiplier),
-		paramtypes.NewParamSetPair(ParamStoreKeyBaseFee, &p.BaseFee, validateBaseFee),
-		paramtypes.NewParamSetPair(ParamStoreKeyEnableHeight, &p.EnableHeight, validateEnableHeight),
-		paramtypes.NewParamSetPair(ParamStoreKeyMinGasPrice, &p.MinGasPrice, validateMinGasPrice),
-		paramtypes.NewParamSetPair(ParamStoreKeyMinGasMultiplier, &p.MinGasMultiplier, validateMinGasPrice),
 	}
 }
 
@@ -119,61 +93,6 @@ func (p Params) Validate() error {
 
 func (p *Params) IsBaseFeeEnabled(height int64) bool {
 	return !p.NoBaseFee && height >= p.EnableHeight
-}
-
-func validateBool(i interface{}) error {
-	_, ok := i.(bool)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-	return nil
-}
-
-func validateBaseFeeChangeDenominator(i interface{}) error {
-	value, ok := i.(uint32)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	if value == 0 {
-		return fmt.Errorf("base fee change denominator cannot be 0")
-	}
-
-	return nil
-}
-
-func validateElasticityMultiplier(i interface{}) error {
-	_, ok := i.(uint32)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-	return nil
-}
-
-func validateBaseFee(i interface{}) error {
-	value, ok := i.(sdkmath.Int)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	if value.IsNegative() {
-		return fmt.Errorf("base fee cannot be negative")
-	}
-
-	return nil
-}
-
-func validateEnableHeight(i interface{}) error {
-	value, ok := i.(int64)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	if value < 0 {
-		return fmt.Errorf("enable height cannot be negative: %d", value)
-	}
-
-	return nil
 }
 
 func validateMinGasPrice(i interface{}) error {